Initial commit to extract initial sync code
