+++ beacon-chain/sync/service.go
@@ -4,7 +4,6 @@ package sync
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/prysmaticlabs/prysm/beacon-chain/types"
 	pb "github.com/prysmaticlabs/prysm/proto/beacon/p2p/v1"
@@ -37,11 +36,6 @@ type Service struct {
 	crystallizedStateBuf         chan p2p.Message
 	announceActiveHashBuf        chan p2p.Message
 	activeStateBuf               chan p2p.Message
-	syncMode                     Mode
-	currentSlotNumber            uint64
-	highestObservedSlot          uint64
-	syncPollingInterval          time.Duration
-	initialCrystallizedStateHash [32]byte
 }
 
 // Config allows the channel's buffer sizes to be changed.
@@ -52,20 +46,7 @@ type Config struct {
 	ActiveStateBufferSize           int
 	CrystallizedStateHashBufferSize int
 	CrystallizedStateBufferSize     int
-	SyncMode                        Mode
-	CurrentSlotNumber               uint64
-	HighestObservedSlot             uint64
-	SyncPollingInterval             time.Duration
 }
-
-// Mode refers to the type for the sync mode of the client.
-type Mode int
-
-// This specifies the different sync modes.
-const (
-	SyncModeInitial Mode = 0
-	SyncModeDefault Mode = 1
-)
 
 // DefaultConfig provides the default configuration for a sync service.
 func DefaultConfig() Config {
@@ -76,26 +57,12 @@ func DefaultConfig() Config {
 		ActiveStateBufferSize:           100,
 		CrystallizedStateHashBufferSize: 100,
 		CrystallizedStateBufferSize:     100,
-		SyncMode:                        SyncModeDefault,
-		CurrentSlotNumber:               0,
-		HighestObservedSlot:             0,
-		SyncPollingInterval:             time.Second,
 	}
 }
 
 // NewSyncService accepts a context and returns a new Service.
 func NewSyncService(ctx context.Context, cfg Config, beaconp2p types.P2P, cs types.ChainService) *Service {
-
 	ctx, cancel := context.WithCancel(ctx)
-	stored, err := cs.HasStoredState()
-
-	if err != nil {
-		log.Errorf("error retrieving stored state: %v", err)
-	}
-
-	if !stored {
-		cfg.SyncMode = SyncModeInitial
-	}
 
 	return &Service{
 		ctx:                          ctx,
@@ -108,24 +75,12 @@ func NewSyncService(ctx context.Context, cfg Config, beaconp2p types.P2P, cs typ
 		crystallizedStateBuf:         make(chan p2p.Message, cfg.ActiveStateBufferSize),
 		announceActiveHashBuf:        make(chan p2p.Message, cfg.CrystallizedStateHashBufferSize),
 		activeStateBuf:               make(chan p2p.Message, cfg.CrystallizedStateBufferSize),
-		syncMode:                     cfg.SyncMode,
-		currentSlotNumber:            cfg.CurrentSlotNumber,
-		highestObservedSlot:          cfg.HighestObservedSlot,
-		syncPollingInterval:          cfg.SyncPollingInterval,
-		initialCrystallizedStateHash: [32]byte{},
 	}
 }
 
 // Start begins the block processing goroutine.
 func (ss *Service) Start() {
-	switch ss.syncMode {
-	case 0:
-		log.Info("Starting initial sync")
-		go ss.runInitialSync(time.NewTicker(ss.syncPollingInterval).C, ss.ctx.Done())
-	default:
-		go ss.run(ss.ctx.Done())
-
-	}
+	go ss.run(ss.ctx.Done())
 }
 
 // Stop kills the block processing goroutine, but does not wait until the goroutine exits.
@@ -246,155 +201,6 @@ func (ss *Service) ReceiveActiveState(data *pb.ActiveState) error {
 	return nil
 }
 
-// RequestCrystallizedStateFromPeer sends a request to a peer for the corresponding crystallized state
-// for a beacon block.
-func (ss *Service) RequestCrystallizedStateFromPeer(data *pb.BeaconBlockResponse, peer p2p.Peer) error {
-	block, err := types.NewBlock(data.Block)
-	if err != nil {
-		return fmt.Errorf("could not instantiate new block from proto: %v", err)
-	}
-	h := block.CrystallizedStateHash()
-	log.Debugf("Successfully processed incoming block with crystallized state hash: %x", h)
-	ss.p2p.Send(&pb.CrystallizedStateRequest{Hash: h[:]}, peer)
-	return nil
-}
-
-// SetBlockForInitialSync sets the first received block as the base finalized
-// block for initial sync.
-func (ss *Service) SetBlockForInitialSync(data *pb.BeaconBlockResponse) error {
-
-	block, err := types.NewBlock(data.Block)
-	if err != nil {
-		return fmt.Errorf("could not instantiate new block from proto: %v", err)
-	}
-
-	h, err := block.Hash()
-	if err != nil {
-		return err
-	}
-	log.WithField("Block received with hash", fmt.Sprintf("0x%x", h)).Debug("Crystallized state hash exists locally")
-
-	if err := ss.writeBlockToDB(block); err != nil {
-		return err
-	}
-
-	ss.initialCrystallizedStateHash = block.CrystallizedStateHash()
-
-	log.Infof("Saved block with hash 0%x for initial sync", h)
-	return nil
-}
-
-// requestNextBlock broadcasts a request for a block with the next slotnumber.
-func (ss *Service) requestNextBlock() {
-	ss.p2p.Broadcast(&pb.BeaconBlockRequestBySlotNumber{SlotNumber: (ss.currentSlotNumber + 1)})
-}
-
-// validateAndSaveNextBlock will validate whether blocks received from the blockfetcher
-// routine can be added to the chain.
-func (ss *Service) validateAndSaveNextBlock(data *pb.BeaconBlockResponse) error {
-	block, err := types.NewBlock(data.Block)
-	if err != nil {
-		return fmt.Errorf("could not instantiate new block from proto: %v", err)
-	}
-
-	if ss.currentSlotNumber == uint64(0) {
-		return fmt.Errorf("invalid slot number for syncing")
-	}
-
-	if (ss.currentSlotNumber + 1) == block.SlotNumber() {
-
-		if err := ss.writeBlockToDB(block); err != nil {
-			return err
-		}
-		ss.currentSlotNumber = block.SlotNumber()
-	}
-	return nil
-}
-
-// writeBlockToDB saves the corresponding block to the local DB.
-func (ss *Service) writeBlockToDB(block *types.Block) error {
-	return ss.chainService.SaveBlock(block)
-}
-
-func (ss *Service) runInitialSync(delaychan <-chan time.Time, done <-chan struct{}) {
-	blockSub := ss.p2p.Subscribe(pb.BeaconBlockResponse{}, ss.blockBuf)
-	crystallizedStateSub := ss.p2p.Subscribe(pb.CrystallizedStateResponse{}, ss.crystallizedStateBuf)
-
-	defer blockSub.Unsubscribe()
-	defer crystallizedStateSub.Unsubscribe()
-	for {
-		select {
-		case <-done:
-			log.Infof("Exiting goroutine")
-			return
-		case <-delaychan:
-			if ss.highestObservedSlot == ss.currentSlotNumber {
-				log.Infof("Exiting initial sync and starting normal sync")
-				go ss.run(ss.ctx.Done())
-				return
-			}
-		case msg := <-ss.blockBuf:
-			data, ok := msg.Data.(*pb.BeaconBlockResponse)
-			// TODO: Handle this at p2p layer.
-			if !ok {
-				log.Errorf("Received malformed beacon block p2p message")
-				continue
-			}
-
-			if data.Block.GetSlotNumber() > ss.highestObservedSlot {
-				ss.highestObservedSlot = data.Block.GetSlotNumber()
-			}
-
-			if ss.currentSlotNumber == 0 {
-				if ss.initialCrystallizedStateHash != [32]byte{} {
-					continue
-				}
-				if err := ss.SetBlockForInitialSync(data); err != nil {
-					log.Errorf("Could not set block for initial sync: %v", err)
-				}
-				if err := ss.RequestCrystallizedStateFromPeer(data, msg.Peer); err != nil {
-					log.Errorf("Could not request crystallized state from peer: %v", err)
-				}
-
-				continue
-			}
-
-			if data.Block.GetSlotNumber() != (ss.currentSlotNumber + 1) {
-				continue
-			}
-
-			if err := ss.validateAndSaveNextBlock(data); err != nil {
-				log.Errorf("Unable to save block: %v", err)
-			}
-			ss.requestNextBlock()
-		case msg := <-ss.crystallizedStateBuf:
-			data, ok := msg.Data.(*pb.CrystallizedStateResponse)
-			// TODO: Handle this at p2p layer.
-			if !ok {
-				log.Errorf("Received malformed crystallized state p2p message")
-				continue
-			}
-
-			if ss.initialCrystallizedStateHash == [32]byte{} {
-				continue
-			}
-
-			crystallizedState := types.NewCrystallizedState(data.CrystallizedState)
-			hash, err := crystallizedState.Hash()
-			if err != nil {
-				log.Errorf("Unable to hash crytsallized state: %v", err)
-			}
-
-			if hash != ss.initialCrystallizedStateHash {
-				continue
-			}
-
-			ss.currentSlotNumber = crystallizedState.LastFinalizedSlot()
-			ss.requestNextBlock()
-			crystallizedStateSub.Unsubscribe()
-		}
-	}
-}
 
 func (ss *Service) run(done <-chan struct{}) {
 	announceBlockHashSub := ss.p2p.Subscribe(pb.BeaconBlockHashAnnounce{}, ss.announceBlockHashBuf)
